backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
sentence-transformers
faiss-cpu
numpy
pandas
pydantic
pydantic-settings
python-dotenv
axios
